requirements.txt
python-binance
python-dotenv
click
colorama
Flask
flask-cors
flask-socketio
google-generativeai
gunicorn
eventlet
